feat: supplier autocomplete

+++ backend/src/Anela.Heblo.Application/Features/Purchase/Model/CreatePurchaseOrderResponse.cs
@@ -6,7 +6,7 @@ public class CreatePurchaseOrderResponse
 {
     public int Id { get; set; }
     public string OrderNumber { get; set; } = null!;
-    public int SupplierId { get; set; }
+    public long SupplierId { get; set; }
     public string SupplierName { get; set; } = null!;
     public DateTime OrderDate { get; set; }
     public DateTime? ExpectedDeliveryDate { get; set; }
